requirements.txt
beautifulsoup4
python-dotenv
streamlit
trafilatura
